requirements.txt
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
biopython==1.84
google-generativeai==0.8.3
tqdm==4.67.0
neo4j==5.28.0
langchain-text-splitters==0.3.6
tiktoken==0.8.0
langchain-huggingface==0.1.2
txtai==8.2.0
cuda-python==12.8.0
langchain==0.3.18
langchain-groq==0.2.4
langchain-google-genai==2.0.11
langchain-ollama==0.2.3
ragas==0.2.14
openai==1.70.0
matplotlib==3.10.1
scipy==1.15.2
torch==2.6.0
torchvision==0.21.0
torch-geometric==2.6.0
seaborn==0.13.2
graphdatascience==1.14